Add big data feature
Adds one function to generate a number of entries and one to summarize
them. The generating function should probably be moved to a backen in
order to not block the browser.

The purpose of this module is to test how transfering a large amount of
data between web workers affects the performance.